polygon-validator-monitor/requirements.txt
aiohttp==3.8.1
web3==5.31.0
pyyaml==6.0
